Expanded documentation Row and Editable Components
Overhauled the documentation of some the editable components and fixed some logic errors that were brough up by the typing check. Also renamed them.
