+++ src/components/DisplayPanel/EditableName.jsx
@@ -1,29 +1,39 @@
 import React, { useEffect, useState } from "react"
 
-const DisplayTaskName = ({ id, name, handleNameUpdate }) => {
+/**
+ * This component will display the name of the Task. When clicked, this component will allow the user to edit the value with an input of type text.
+ * @param {Object} props - Component props object
+ * @param {Number} props.id - The id of the Task associated to this component object contained in this row, used when setting the id for the input tag.
+ * @param {String} props.name - The name the Task already has.
+ * @param {function(String):void} props.handleNameUpdate - Callback function that will be fired when the changes are submitted.
+ */
+const EditableName = ({ id, name, handleNameUpdate }) => {
 
+    /** This will be used to control the input tag. */
     const [tempName, setTempName] = useState(name)
+    /** This value controls the conditional rendering for when the component is in editing mode. */
     const [isEditingName, setIsEditingName] = useState(false)
-    
+
+    /** This will be used to control the input tag. */
     useEffect(() => setTempName(name), [name])
-    
+
     const handleSubmit = event => {
         event.preventDefault()
-        // console.log(`An edit for the name has been submitted for the task ${id}`)
-        // Form validations
+        //      Form Validation
         if (tempName === name) {
+            // If the name is the same don't even update the state
             setIsEditingName(false)
             return
         }
         if (tempName === "") {
+            // Do not allow empty names
             alert("The name cannot be empty") // #TODO Implement a better notification
             setTempName(name)
             return
         }
-        // Edit submission
-        // console.log("About to fire the handleNameUpdate")
+        //      Update Task
         handleNameUpdate(tempName)
-        // Cleanup
+        //      Cleanup
         setIsEditingName(false)
     }
 
@@ -43,22 +53,14 @@ const DisplayTaskName = ({ id, name, handleNameUpdate }) => {
                         />
                     </form>
                     :
-                    <button
-                        onClick={() => setIsEditingName(true)}
-                    >
+                    <button onClick={() => setIsEditingName(true)}>
                         <span>
                             {name}
                         </span>
                     </button>
             }
-            {/* <button
-                onClick={() => setIsEditingName(true)}
-                disabled={isEditingName}
-            >
-                Editar nombre
-            </button> */}
         </>
     )
 }
 
-export default DisplayTaskName;
+export default EditableName;
